pyproject.toml
[tool.poetry]
name = "agrivision-article-processor"
version = "0.1.0"
description = "Traitement d'articles scientifiques avec LangChain et Pydantic"
authors = ["Votre Nom <votre.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-community = "^0.0.20"
langchain-mistralai = "^0.1.0"
pydantic = "^2.0.0"
unstructured = {extras = ["md"], version = "^0.10.0"}
python-magic-bin = "^0.4.14"
python-dotenv = "^1.0.0"
pandas = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pre-commit = "^3.0.0"

[tool.poetry.scripts]
process-articles = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true